+++ kite/src/main/kotlin/com/tang/kite/wrapper/Wrapper.kt
@@ -11,4 +11,6 @@ interface Wrapper<T> {
 
     fun getSqlStatement(): SqlStatement
 
+    fun checkValues()
+
 }