+++ kite/src/main/kotlin/com/tang/kite/wrapper/delete/DeleteWrapper.kt
@@ -1,26 +1,24 @@
-package com.tang.kite.wrapper.update
+package com.tang.kite.wrapper.delete
 
 import com.tang.kite.config.SqlConfig
-import com.tang.kite.constants.SqlString.UPDATE
+import com.tang.kite.constants.SqlString.DELETE_FROM
 import com.tang.kite.mapper.BaseMapper
 import com.tang.kite.sql.SqlStatement
 import com.tang.kite.utils.Reflects
 import com.tang.kite.wrapper.Wrapper
 
 /**
- * Update wrapper for update operation
+ * Delete wrapper for delete operation
  *
  * @author Tang
  */
-class UpdateWrapper<T> : Wrapper<T> {
+class DeleteWrapper<T> : Wrapper<T> {
 
     private lateinit var table: String
 
     lateinit var baseMapper: BaseMapper<T>
 
-    private lateinit var updateSetWrapper: UpdateSetWrapper<T>
-
-    lateinit var updateWhereWrapper: UpdateWhereWrapper<T>
+    lateinit var deleteWhereWrapper: DeleteWhereWrapper<T>
 
     constructor()
 
@@ -31,15 +29,14 @@ class UpdateWrapper<T> : Wrapper<T> {
     companion object {
 
         /**
-         * Create a new UpdateWrapper instance
+         * Create a new DeleteWrapper instance
          *
-         * @return UpdateWrapper
+         * @return DeleteWrapper
          */
         @JvmStatic
-        fun <T> create(): UpdateWrapper<T> {
-            return UpdateWrapper()
+        fun <T> create(): DeleteWrapper<T> {
+            return DeleteWrapper()
         }
-
     }
 
     /**
@@ -47,10 +44,10 @@ class UpdateWrapper<T> : Wrapper<T> {
      *
      * @param table table name
      */
-    fun from(table: String): UpdateSetWrapper<T> {
+    fun from(table: String): DeleteWhereWrapper<T> {
         this.table = table
-        this.updateSetWrapper = UpdateSetWrapper(this)
-        return updateSetWrapper
+        this.deleteWhereWrapper = DeleteWhereWrapper(this)
+        return deleteWhereWrapper
     }
 
     /**
@@ -58,22 +55,16 @@ class UpdateWrapper<T> : Wrapper<T> {
      *
      * @param clazz entity class
      */
-    fun from(clazz: Class<T>): UpdateSetWrapper<T> {
+    fun from(clazz: Class<T>): DeleteWhereWrapper<T> {
         return from(Reflects.getTableName(clazz))
     }
 
-    /**
-     * Get the SQL statement
-     *
-     * @return SqlStatement
-     */
     override fun getSqlStatement(): SqlStatement {
         checkValues()
         val sql: StringBuilder = StringBuilder()
         val parameters: MutableList<Any?> = mutableListOf()
-        sql.append("$UPDATE$table")
-        updateSetWrapper.appendSql(sql, parameters)
-        updateWhereWrapper.appendSql(sql, parameters)
+        sql.append("$DELETE_FROM$table")
+        deleteWhereWrapper.appendSql(sql, parameters)
         return SqlStatement(SqlConfig.getSql(sql), parameters)
     }