+++ kite/src/main/kotlin/com/tang/kite/wrapper/query/QueryWrapper.kt
@@ -117,7 +117,7 @@ class QueryWrapper<T> : Wrapper<T> {
         return SqlStatement(SqlConfig.getSql(sql), parameters)
     }
 
-    private fun checkValues() {
+    override fun checkValues() {
     }
 
 }